test/e2e: refactor e2e tests
